requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
requests
openai
httpx
Jinja2
semantic-kernel
